created Offers component
